Refactored to include all projects.
