+++ common/inference.py
@@ -1,0 +1,118 @@
+"""
+ Copyright (c) 2018 Intel Corporation.
+
+ Permission is hereby granted, free of charge, to any person obtaining
+ a copy of this software and associated documentation files (the
+ "Software"), to deal in the Software without restriction, including
+ without limitation the rights to use, copy, modify, merge, publish,
+ distribute, sublicense, and/or sell copies of the Software, and to
+ permit persons to whom the Software is furnished to do so, subject to
+ the following conditions:
+
+ The above copyright notice and this permission notice shall be
+ included in all copies or substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+"""
+
+import os
+import sys
+import logging as log
+from openvino.inference_engine import IENetwork, IECore
+from datetime import datetime
+
+
+class Network:
+    """
+    Load and configure inference plugins for the specified target devices
+    and performs synchronous and asynchronous modes for the specified infer requests.
+    """
+
+    def __init__(self):
+        ### TODO: Initialize any class variables desired ###
+        self.plugin = None
+        self.network = None
+        self.input_blob = None
+        self.output_blob = None
+        self.exec_network = None
+        self.infer_request = None
+
+        log.basicConfig(format="[ %(levelname)s ] %(message)s", level=log.INFO, stream=sys.stdout)
+
+
+    def load_model(self, model, dev="CPU", cpu_extension=None, name="inc"):
+        ### TODO: Load the model ###
+        ### TODO: Check for supported layers ###
+        ### TODO: Add any necessary extensions ###
+        ### TODO: Return the loaded inference plugin ###
+        ### Note: You may need to update the function parameters. ###
+        model_xml = model
+        model_bin = os.path.splitext(model_xml)[0] + ".bin"
+
+        # Initialize the plugin
+        self.plugin = IECore()
+
+        # Add a CPU extension, if applicable
+        if cpu_extension and "CPU" in dev:
+            self.plugin.add_extension(cpu_extension, dev)
+
+        # Read the IR as a IENetwork
+        #log.info("Loading network files:\n\t{}\n\t{}".format(model_xml, model_bin))
+        self.network = IENetwork(model=model_xml, weights=model_bin)
+
+        # Get the supported layers of the network
+        print("Infer Device: ", dev)
+        supported_layers = self.plugin.query_network(network=self.network, device_name=dev)
+
+        # Check for any unsupported layers, and let the user
+        # know if anything is missing. Exit the program, if so.
+        if "CPU" in dev:
+            unsupported_layers = [l for l in self.network.layers.keys() if l not in supported_layers]
+            if len(unsupported_layers) != 0:
+                log.error("The following layers are not supported by the plugin for specified device {}:\n {}".format(dev, ', '.join(unsupported_layers)))
+                log.error("Please try to specify cpu extensions library path in sample's command line parameters using -l or --cpu_extension command line argument")
+                sys.exit(1)
+
+        # Load the IENetwork into the plugin
+        start_time = datetime.now()
+        self.exec_network = self.plugin.load_network(self.network, dev)
+        duration_ms = (datetime.now() - start_time).total_seconds() * 1000
+
+        # Get the input layer
+        if name == 'rcnn':
+            self.input_blob = 'image_tensor'
+        else:
+            self.input_blob = next(iter(self.network.inputs))
+
+        self.output_blob = next(iter(self.network.outputs))
+
+        return duration_ms
+
+    def get_input_shape(self):
+        ### TODO: Return the shape of the input layer ###
+        return self.network.inputs[self.input_blob].shape
+
+    def exec_net(self, image):
+        ### TODO: Start an asynchronous request ###
+        ### TODO: Return any necessary information ###
+        ### Note: You may need to update the function parameters. ###
+        infer_request_handle = self.exec_network.start_async(request_id=0,inputs={self.input_blob: image})
+        return
+
+    def wait(self):
+        ### TODO: Wait for the request to be complete. ###
+        ### TODO: Return any necessary information ###
+        ### Note: You may need to update the function parameters. ###
+        status = self.exec_network.requests[0].wait()
+        return status
+
+    def get_output(self):
+        ### TODO: Extract and return the output results
+        ### Note: You may need to update the function parameters. ###
+        return self.exec_network.requests[0].outputs[self.output_blob]
