+++ models/intel/ssd_mobilenet_v2_coco/ssd_mobilenet_v2_coco_2018_03_29/checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "model.ckpt"
+all_model_checkpoint_paths: "model.ckpt"
