+++ move-mouse-pointer/mouse_controller.py
@@ -1,0 +1,28 @@
+'''
+This is a sample class that you can use to control the mouse pointer.
+It uses the pyautogui library. You can set the precision for mouse movement
+(how much the mouse moves) and the speed (how fast it moves) by changing
+precision_dict and speed_dict.
+Calling the move function with the x and y output of the gaze estimation model
+will move the pointer.
+This class is provided to help get you started; you can choose whether you want to use it or create your own from scratch.
+'''
+import pyautogui
+
+class MouseController:
+    def __init__(self, precision, speed):
+        precision_dict={'high':100, 'low':1000, 'medium':500}
+        speed_dict={'fast':1, 'slow':10, 'medium':5}
+
+        self.precision=precision_dict[precision]
+        self.speed=speed_dict[speed]
+        pyautogui.FAILSAFE = False
+
+    def move(self, x, y):
+        pyautogui.moveRel(x*self.precision, -1*y*self.precision, duration=self.speed)
+
+    def monitor(self):
+        return pyautogui.size()
+
+    def put(self, x, y):
+        pyautogui.moveTo(x, y)
