+++ move-mouse-pointer/runcam.bat
@@ -1,0 +1,2 @@
+::"C:\Program Files (x86)\IntelSWTools\openvino_2020.1.033\bin\setupvars.bat"
+python main_local.py -i 0 -fdm C:\Users\jlgarci2\Documents\Intel\OpenVINO_2019\openvino_models\intel\face-detection-adas-binary-0001\FP32-INT1\face-detection-adas-binary-0001 -flm C:\Users\jlgarci2\Documents\Intel\OpenVINO_2019\openvino_models\intel\landmarks-regression-retail-0009\FP16\landmarks-regression-retail-0009 -hpm C:\Users\jlgarci2\Documents\Intel\OpenVINO_2019\openvino_models\intel\head-pose-estimation-adas-0001\FP16\head-pose-estimation-adas-0001 -gem C:\Users\jlgarci2\Documents\Intel\OpenVINO_2019\openvino_models\intel\gaze-estimation-adas-0002\FP16\gaze-estimation-adas-0002 -d CPU -pt 0.3
