--- src/rundemo.bat
@@ -1,8 +1,0 @@
-python main.py -nf 600 -p FP32,FP16 -i ..\bin\demo.mp4 -fdm ..\models\intel\face-detection-adas-binary-0001\FP32-INT1\face-detection-adas-binary-0001 -flm ..\models\intel\landmarks-regression-retail-0009 -hpm ..\models\intel\head-pose-estimation-adas-0001 -gem ..\models\intel\gaze-estimation-adas-0002 -d CPU -ct 0.3 -async False
-::python main.py -nf 5 -p FP16 -i ..\bin\demo.mp4 -fdm ..\models\intel\face-detection-adas-binary-0001\FP32-INT1\face-detection-adas-binary-0001 -flm ..\models\intel\landmarks-regression-retail-0009 -hpm ..\models\intel\head-pose-estimation-adas-0001 -gem ..\models\intel\gaze-estimation-adas-0002 -d HETERO:GPU,CPU -ct 0.3
-
-::python main.py -fancy True -p FP16 -bm True -i 0 -fdm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\face-detection-adas-binary-0001\FP32-INT1\face-detection-adas-binary-0001 -flm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\landmarks-regression-retail-0009\FP16\landmarks-regression-retail-0009 -hpm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\head-pose-estimation-adas-0001\FP16\head-pose-estimation-adas-0001 -gem C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\gaze-estimation-adas-0002\FP16\gaze-estimation-adas-0002 -d CPU -ct 0.3
-
-::python main.py -fancy True -p FP16 -bm True -i "C:\Users\jlgarci2\Dropbox\github\move-mouse-pointer\bin\demo.mp4" -fdm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\face-detection-adas-binary-0001\FP32-INT1\face-detection-adas-binary-0001 -flm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\landmarks-regression-retail-0009\FP16\landmarks-regression-retail-0009 -hpm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\head-pose-estimation-adas-0001\FP16\head-pose-estimation-adas-0001 -gem C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\gaze-estimation-adas-0002\FP16\gaze-estimation-adas-0002 -d HETERO:GPU,CPU -ct 0.3
-
-::python main.py -fancy True -p FP16 -bm True -i "C:\Users\jlgarci2\Dropbox\github\move-mouse-pointer\bin\demo.mp4" -fdm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\face-detection-adas-binary-0001\FP32-INT1\face-detection-adas-binary-0001 -flm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\landmarks-regression-retail-0009\FP16\landmarks-regression-retail-0009 -hpm C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\head-pose-estimation-adas-0001\FP16\head-pose-estimation-adas-0001 -gem C:\Users\jlgarci2\Documents\Intel\OpenVINO\openvino_models\intel\gaze-estimation-adas-0002\FP16\gaze-estimation-adas-0002 -d HETERO:MYRIAD,CPU -ct 0.3
